+++ extensions/cursor-recent-projects/CHANGELOG.md
@@ -1,4 +1,15 @@
-# Cursor Search Recent Projects Changelog
+# Cursor Changelog
+
+## [Documentations] - 2024-10-15
+
+- Correct extension title in readme and changelog
+- Add extension slogan below the title
+- Add a badge for the Raycast Cross-Extension Conventions
+- Add documentation for the Cross-Extension usage
+
+## [Features] - 2024-10-11
+
+- Add integration with [Cursor Directory](https://www.raycast.com/escwxyz/cursor-directory)
 
 ## [Initial Version] - 2024-09-25